+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/notifications/data/NotificationsPreferencesRepositoryImpl.kt
@@ -134,7 +134,7 @@ internal class NotificationsPreferencesRepositoryImpl @Inject constructor(
                 notificationCount,
             )
         }
-    }.take(1).firstOrNull()
+    }.firstOrNull()
 
     private fun getActionsSummary(): String {
         val userActions = settings.newEpisodeNotificationActions.value
